requirements.txt
streamlit>=1.25.0
scikit-learn==1.6.1   # This is the version from your Colab output
numpy>=1.23.0
PyPDF2>=3.0.1
python-docx>=0.8.11
matplotlib>=3.5.0
reportlab>=3.6.0
joblib==1.5.1         # This is the version from your Colab output
pandas==2.2.2         # This is the version from your Colab output
nltk>=3.6.0
seaborn>=0.11.0